rename: ShuttleRepository 위치 변경
