+++ seurasaeng_be/src/main/java/onehajo/seurasaeng/qr/service/QRService.java
@@ -12,7 +12,7 @@ import onehajo.seurasaeng.qr.dto.ValidUserResDTO;
 import onehajo.seurasaeng.qr.exception.InvalidQRCodeException;
 import onehajo.seurasaeng.qr.exception.UserNotFoundException;
 import onehajo.seurasaeng.qr.repository.QrRepository;
-import onehajo.seurasaeng.qr.repository.shuttle.ShuttleRepository;
+import onehajo.seurasaeng.shuttle.repository.ShuttleRepository;
 import onehajo.seurasaeng.user.repository.UserRepository;
 import onehajo.seurasaeng.qr.util.AESUtil;
 import onehajo.seurasaeng.entity.Boarding;
